+++ code/pyorg/surf/utils.py
@@ -356,19 +356,31 @@ def points_to_polyline(coords):
     poly.SetLines(cells)
     return poly
 
-def polyline_to_points(polyline):
+def polyline_to_points(polyline, mode='verts'):
     """
     Extract an ordered list with the points in a polyline
     :param polyline: input vtkPolyData with a polyline
+    :param mode: valid; 'verts' cell vertices, 'line' cell line
     :return: numpy array with point coordinates ordered
     """
-    verts = polyline.GetVerts()
-    coords = np.zeros(shape=(verts.GetNumberOfCells(), 3), dtype=np.float32)
-    for vert_id in range(verts.GetNumberOfCells()):
-        cell = vtk.vtkGenericCell()
-        verts.GetCell(vert_id, cell)
-        pts = cell.GetPoints()
-        coords[vert_id, :] = pts.GetPoint(0)
+    assert (mode == 'verts') or (mode == 'line')
+    if mode == 'verts':
+        verts = polyline.GetVerts()
+        coords = np.zeros(shape=(verts.GetNumberOfCells(), 3), dtype=np.float32)
+        for vert_id in range(verts.GetNumberOfCells()):
+            cell = vtk.vtkGenericCell()
+            verts.GetCell(vert_id, cell)
+            pts = cell.GetPoints()
+            coords[vert_id, :] = pts.GetPoint(0)
+    elif mode == 'line':
+        lines = polyline.GetLines()
+        coords = list()
+        for line_id in range(lines.GetNumberOfCells()):
+            ids = vtk.vtkIdList()
+            lines.GetCell(line_id, ids)
+            for i in range(ids.GetNumberOfIds()):
+                coords.append(polyline.GetPoint(ids.GetId(i)))
+        coords = np.asarray(coords, dtype=np.float32)
     return coords
 
 def is_point_inside_surf(point, selector, conv_iter, max_iter):